added blob_sidecar_list_indexed to block_id
